+++ src/components/inputs/Input.tsx
@@ -1,10 +1,8 @@
-import React, { ReactNode, useState } from 'react';
-import DatePicker from 'react-datepicker';
+import React from 'react';
 import 'react-datepicker/dist/react-datepicker.css';
-import TagChip from '@/components/chips/TagChip';
-import { IconCalendar, IconGreater, IconLess } from '@/public/svgs';
+import { InputComponents } from '@/components/inputs/InputComponents';
 
-type Props = {
+export type Props = {
   title?: string;
   required?: boolean;
   value?: string | number;
@@ -25,186 +23,6 @@ function Label({ title, required }: Props) {
   );
 }
 
-function TextInput({
-  required,
-  children,
-  ...rest
-}: {
-  required: boolean;
-  children: ReactNode;
-}) {
-  return (
-    <input className='input' required={required} {...rest}>
-      {children}
-    </input>
-  );
-}
-
-function Textarea({
-  required,
-  children,
-  ...rest
-}: {
-  required: boolean;
-  children: ReactNode;
-}) {
-  return (
-    <textarea className='textarea' required={required} {...rest}>
-      {children}
-    </textarea>
-  );
-}
-
-function DateInput({ required }: { required: boolean }) {
-  const handleChange = (e: React.ChangeEvent) => {
-    e.preventDefault();
-  };
-  const [value, setValue]: [Date | null, Function] = useState(null);
-  const YEARS = Array.from(
-    { length: new Date().getFullYear() + 1 - 2000 },
-    (_, i) => new Date().getFullYear() - i,
-  );
-  const MONTHS = [
-    'January',
-    'February',
-    'March',
-    'April',
-    'May',
-    'June',
-    'July',
-    'August',
-    'September',
-    'October',
-    'November',
-    'December',
-  ];
-  return (
-    <DatePicker
-      renderCustomHeader={({
-        date,
-        changeYear,
-        decreaseMonth,
-        increaseMonth,
-        prevMonthButtonDisabled,
-        nextMonthButtonDisabled,
-      }) => (
-        <div className='flex justify-between'>
-          <div className='custom-header'>
-            <span className='ml-12 font-bold'>{MONTHS[date.getMonth()]}</span>
-            <select
-              value={date.getFullYear()}
-              className='font-bold outline-none'
-              onChange={({ target: { value } }) => changeYear(+value)}
-            >
-              {YEARS.map((option) => (
-                <option key={option} value={option}>
-                  {option}
-                </option>
-              ))}
-            </select>
-          </div>
-          <div className='mr-12 flex gap-12'>
-            <button
-              type='button'
-              onClick={decreaseMonth}
-              disabled={prevMonthButtonDisabled}
-            >
-              <IconLess className='h-12 w-12' />
-            </button>
-            <button
-              type='button'
-              onClick={increaseMonth}
-              disabled={nextMonthButtonDisabled}
-            >
-              <IconGreater className='h-12 w-12' />
-            </button>
-          </div>
-        </div>
-      )}
-      showYearDropdown
-      scrollableYearDropdown
-      dateFormat='yyyy.MM.dd HH:mm'
-      dateFormatCalendar='yyyy년 MM월'
-      showTimeSelect
-      timeFormat='HH:mm'
-      timeIntervals={30}
-      timeCaption='시간'
-      showPopperArrow={false}
-      fixedHeight
-      className='input'
-      placeholderText='날짜를 입력해 주세요'
-      showIcon
-      selected={value}
-      onChange={(date: Date) => setValue(date)}
-      icon={<IconCalendar />}
-      onChangeRaw={handleChange}
-      required={required}
-    />
-  );
-}
-
-function TagInput({
-  children,
-  tagList = [],
-  setTagList = () => {},
-  ...rest
-}: Props) {
-  const [tagItem, setTagItem] = useState<string>('');
-
-  const handleTagValue = (e: React.KeyboardEvent<HTMLInputElement>) => {
-    const target = e.target as HTMLInputElement;
-    if (e.key === 'Enter' && target.value.length !== 0) {
-      addTagItem();
-      return;
-    }
-    if (e.key === 'Backspace' && tagList?.length !== 0) {
-      deleteTagItem(tagList.length - 1);
-    }
-  };
-
-  const addTagItem = () => {
-    if (tagItem.trim() !== '') {
-      setTagList([...tagList, tagItem]);
-      setTagItem('');
-    }
-  };
-
-  const deleteTagItem = (index: number) => {
-    const filteredTagList = [...tagList];
-    filteredTagList.splice(index, 1);
-    setTagList(filteredTagList);
-  };
-
-  return (
-    <div className='input flex h-auto items-center'>
-      <div className='flex w-full flex-wrap  gap-8 '>
-        {tagList.map((tag, index) => (
-          <div className='flex shrink-0' key={index}>
-            <TagChip str={tag}>
-              <button
-                type='button'
-                className='text-8 ml-5 text-gray-5 tablet:h-20 tablet:text-10'
-                onClick={() => deleteTagItem(index)}
-              >
-                x
-              </button>
-            </TagChip>
-          </div>
-        ))}
-        <input
-          className='input-no-style'
-          {...rest}
-          onKeyUp={handleTagValue}
-          onChange={(e) => setTagItem(e.target.value)}
-          value={tagItem}
-        >
-          {children}
-        </input>
-      </div>
-    </div>
-  );
-}
-
 function SelectedInput({
   type,
   required = false,
@@ -216,23 +34,27 @@ function SelectedInput({
   switch (type) {
     case 'text':
       return (
-        <TextInput required={required} {...rest}>
+        <InputComponents.TextInput required={required} {...rest}>
           {children}
-        </TextInput>
+        </InputComponents.TextInput>
       );
     case 'date':
-      return <DateInput required={required} />;
+      return <InputComponents.DateInput required={required} />;
     case 'textarea':
       return (
-        <Textarea required={required} {...rest}>
+        <InputComponents.Textarea required={required} {...rest}>
           {children}
-        </Textarea>
+        </InputComponents.Textarea>
       );
     case 'tag':
       return (
-        <TagInput tagList={tagList} setTagList={setTagList} {...rest}>
+        <InputComponents.TagInput
+          tagList={tagList}
+          setTagList={setTagList}
+          {...rest}
+        >
           {children}
-        </TagInput>
+        </InputComponents.TagInput>
       );
   }
 }
