+++ src/components/inputs/InputComponents/Textarea.tsx
@@ -1,0 +1,18 @@
+import React, { ReactNode } from 'react';
+
+function Textarea({
+  required,
+  children,
+  ...rest
+}: {
+  required: boolean;
+  children: ReactNode;
+}) {
+  return (
+    <textarea className='textarea' required={required} {...rest}>
+      {children}
+    </textarea>
+  );
+}
+
+export default Textarea;
